code cleanup, add network thread
